Move memoization section into advanced callbacks chapter.
